Make acktr's atari do more than create a model and throw it away
Alter acktr's run_atari.py to save its model at least at the end of
training, and optionally every X updates.
Add an enjoy_atari.py to use the trained model to display the game being
played.
Alter acktr_disc.py to be able to resume training a model by copying a
model file to the parent directory of the logger and calling it
"checkpoint_model".  The counts are not restored, just the previously
trained attributes of the model.